requirements.txt
fpdf==1.7.2
Pillow==6.2.2
PyPDF2==1.26.0